+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.5.0-SNAPSHOT] - 2025-03-4
+
+### Changed
+
+- now using nanoTime
+- improved calculateLcmOfRange by introducing nextPower
+- improved sieveOfEratosthenes method by generating boolean array of composite numbers
+
 ## [1.4.0-SNAPSHOT] - 2025-03-19
 
 ### Added